Moved returnFirstXWords to common folder. Replacing "beta" under logo with "ballot".
